requirements.txt
flask
python-dotenv
openai
aiohttp
requests
langchain_openai
#SQLAlchemy
langchain-community
pymysql
cryptography
faiss-cpu
gunicorn
google-cloud-storage
python-telegram-bot
google-cloud-aiplatform
langchain-google-community[gmail]
langchain[openai]
langgraph
SQLAlchemy==2.0.29
xhtml2pdf
python-docx